+++ src/index.ts
@@ -1,3 +1,3 @@
 export { BigWig } from './bigwig'
 export { BigBed } from './bigbed'
-export { Feature, Header } from './bbi'
+export { Feature, Header, RequestOptions } from './bbi'